pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blast-summary-tool"
version = "0.1.0"
authors = [
    {name = "BLAST Summary Tool Developer"}
]
description = "AI-powered analysis and summarization of BLAST results"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
blast-summary = "blast_summary.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/blast-summary-tool"

[tool.setuptools.packages.find]
where = ["."]
include = ["blast_summary*"]
